feat: move hashing from cat-type to cbc-type
