+++ rust/cardano-blockchain-types/Cargo.toml
@@ -38,3 +38,5 @@ ed25519-dalek = "2.1.1"
 serde = "1.0.210"
 num-bigint = "0.4.6"
 serde_json = "1.0.134"
+displaydoc = "0.2.5"
+thiserror = "2.0.11"
