+++ rust/cardano-blockchain-types/src/hashes/hash_wrapper.rs
